Support non-fixed dates for the last week of the month
Princess cake week is a whole week, where the whole week is in September
(i.e. Sunday is in September). The alternative could be the last week
starting with the last Monday of the month, but this blog hints at the
former:
https://prvbloggen.se/2015/09/21/prinsesstartan-en-generisk-framgang-som-denna-vecka-bidrar-till-forskning/

E.g. in 2020 the Princess cake week is 21/9 - 27/9 and not 28/9 - 4/10.